Embind only accepts int and double.

+++ nui/include/nui/frontend/dom/childless_element.hpp
@@ -53,13 +53,13 @@ namespace Nui::Dom
         requires std::integral<T>
         void setAttribute(std::string_view key, T value)
         {
-            element_.call<Nui::val>("setAttribute", Nui::val{std::string{key}}, Nui::val{value});
+            element_.call<Nui::val>("setAttribute", Nui::val{std::string{key}}, Nui::val{static_cast<int>(value)});
         }
         template <typename T>
         requires std::floating_point<T>
         void setAttribute(std::string_view key, T value)
         {
-            element_.call<Nui::val>("setAttribute", Nui::val{std::string{key}}, Nui::val{value});
+            element_.call<Nui::val>("setAttribute", Nui::val{std::string{key}}, Nui::val{static_cast<double>(value)});
         }
         void setAttribute(std::string_view key, Nui::val value)
         {
